# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Core Swagger",
        "description": "Swagger for streamfiuses core api",
        "version": "1.0.0"
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "paths": {
...
        "/api/content/{id}  ": {
            "delete": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "tags": [
                    "Content"
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "type": "integer",
                        "required": true,
                        "description": "Numeric ID of the content to delete."
                    }
                ],
                "description": "Delete a specific content",
                "operationId": "App\\Http\\Controllers\\ContentController::destroy",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "unauthorized"
                    },
                    "404": {
                        "description": "Content not found"
                    },
                    "500": {
                        "description": "server error"
                    }
                }
            }
        },
        "/api/content/multiple/{content_ids}  ": {
            "get": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "tags": [
                    "Content"
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "content_ids",
                        "type": "string",
                        "required": true,
                        "description": "Json array string of content ids to fetch"
                    }
                ],
                "description": "Delete a specific content",
                "operationId": "App\\Http\\Controllers\\ContentController::destroy",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "unauthorized"
                    },
                    "404": {
                        "description": "Content not found"
                    },
                    "500": {
                        "description": "server error"
                    }
                }
            }
        },
...
    },
...
}